==> Análisis y Cálculo de Calidad de vida/1. Comercios.ipynb <==
# Análisis de actividades económicas
## 1. Importar librerías y establecer el directorio de trabajo

import os
import geopandas as gpd

os.chdir("C:/Users/aleja/OneDrive - Universidad Politécnica de Madrid/Universidad/Asignaturas/Cuarto/Trabajo de Fin de Grado")

## 2. Cargar datos

actividades_economicas = gpd.read_file("Capas/ACTIVIDADES.geojson")
actividades_economicas.head()

## 3. Limpieza de datos
Se eliminan las categorías duplicadas para saber qué comercios hay

comercios = actividades_economicas.drop_duplicates(['categoria'], keep='first')
comercios.reset_index(drop=True, inplace=True)
print(comercios['categoria'])

## 4. Guardar datos
Una vez ya se saben las categorías que hay, se almacenarán los datos en un excel para facilitar el visionado y la manipulación de estos en un futuro

comercios.to_excel('Resultados/comercios.xlsx')
