==> Análisis y Cálculo de Calidad de vida/2. Análisis y cálculo de distancias a accesos.ipynb <==
# Análisis de datos y cálculo de distancia euclídea
## 1. Importación de librerías y establecimiento de directorio de trabajo

import os
import geopandas as gpd
import requests
import urllib.parse
import warnings
warnings.simplefilter("ignore")
import time
import random
import haversine as hs

os.chdir("C:/Users/aleja/OneDrive - Universidad Politécnica de Madrid/Universidad/Asignaturas/Cuarto/Trabajo de Fin de Grado/")

## 2. Carga de datos

accesos = gpd.read_file("Capas/ACCESOS.geojson")
actividades_economicas = gpd.read_file("Capas/ACTIVIDADES.geojson")
administracion = gpd.read_file("Capas/POIS/poi_administracio_ca.geojson")
cultura = gpd.read_file("Capas/POIS/poi_cultura_ca.geojson")
educacion = gpd.read_file("Capas/POIS/poi_educacio_ca.geojson")
deporte = gpd.read_file("Capas/POIS/poi_esports_ca.geojson")
ocio = gpd.read_file("Capas/POIS/poi_lloi_ca.geojson")
sanidad = gpd.read_file("Capas/POIS/poi_sanitat_ca.geojson")
servicios = gpd.read_file("Capas/POIS/poi_serveis_ca.geojson")
movilidad = gpd.read_file("Capas/POIS/poi_transports_mobilitat_ca.geojson")

## 3. Limpieza y visualización de datos
Se hará una visualización de todos los datos y se agruparán por indicador

Todos deberán estar en UTM 31

También se hará una limpieza de los datos

Solo se guardará la latitud y la longitud que se obtendrán del atributo geometry

### 3.0. Accesos

accesos.head()

accesos["latitud"] = accesos["geometry"].y
accesos["longitud"] = accesos["geometry"].x
accesos = accesos.drop(columns=['id', 'num_acceso', 'refcat', 'geometry'])
accesos.head()

### 3.1. Actividades económicas

actividades_economicas.head()

# se eliminan los atributos que innecesarios
actividades_economicas["latitud"] = actividades_economicas["geometry"].y
actividades_economicas["longitud"] = actividades_economicas["geometry"].x
actividades_economicas = actividades_economicas.drop(columns=['id', 'aa_ida', 'geometry'])
actividades_economicas.head()

#### 3.1.1. Comercios básicos

pescaderias = actividades_economicas[actividades_economicas["categoria"] == "106###PEIXETERIA"]
pescaderias = pescaderias.drop(columns=["categoria"])
pescaderias.reset_index(drop=True, inplace=True)

supermercados = actividades_economicas[actividades_economicas["categoria"] == "129###SUPERMERCATS/HIPERMERCATS"]
supermercados = supermercados.drop(columns=["categoria"])
supermercados.reset_index(drop=True, inplace=True)

fruterias = actividades_economicas[actividades_economicas["categoria"] == "73###FRUITA I VERDURA"]
fruterias = fruterias.drop(columns=["categoria"])
fruterias.reset_index(drop=True, inplace=True)

alimentacion = actividades_economicas[actividades_economicas["categoria"] == "12###ALIMENTACIÓ, BEGUDES I TABAC"]
alimentacion = alimentacion.drop(columns=["categoria"])
alimentacion.reset_index(drop=True, inplace=True)

panaderias = actividades_economicas[actividades_economicas["categoria"] == "71###FORNS I PASTISSERIES"]
panaderias = panaderias.drop(columns=["categoria"])
panaderias.reset_index(drop=True, inplace=True)

carnicerias = actividades_economicas[actividades_economicas["categoria"] == "37###CARNISSERIA"]
carnicerias = carnicerias.drop(columns=["categoria"])
carnicerias.reset_index(drop=True, inplace=True)

alimentacion_no_especializada = actividades_economicas[
    actividades_economicas["categoria"] == "34###BOTIGA D'ALIMENTACIÓ NO ESPECIALITZADA"]
alimentacion_no_especializada = alimentacion_no_especializada.drop(columns=["categoria"])
alimentacion_no_especializada.reset_index(drop=True, inplace=True)

peluquerias = actividades_economicas[actividades_economicas["categoria"] == "110###PERRUQUERIES"]
peluquerias = peluquerias.drop(columns=["categoria"])
peluquerias.reset_index(drop=True, inplace=True)

servicios["latitud"] = servicios["geometry"].y
servicios["longitud"] = servicios["geometry"].x
servicios = servicios.drop(columns=["f_id", "poi_category", "poi_type", "poi_adreca", "poi_telefon", "poi_web", "poi_horari", 
                                "poi_observacions", "tipo_poi", "categoria", "idioma", "atributo", "geometry"])

mercado_municipal = servicios[servicios["poi_nom"] == "Mercat municipal"]
mercado_municipal = mercado_municipal.drop(columns=["poi_nom"])
mercado_municipal.reset_index(drop=True, inplace=True)

mercado_ambulante = servicios[servicios["poi_nom"] == "Mercat ambulant"]
mercado_ambulante = mercado_ambulante.drop(columns=["poi_nom"])
mercado_ambulante.reset_index(drop=True, inplace=True)

#### 3.1.2 Comercios secundarios

electricistas_fontaneros = actividades_economicas[actividades_economicas["categoria"] == "58###ELECTRICISTA I LAMPISTA"]
electricistas_fontaneros = electricistas_fontaneros.drop(columns=["categoria"])
electricistas_fontaneros.reset_index(drop=True, inplace=True)

ferreterias = actividades_economicas[actividades_economicas["categoria"] == "68###FERRETERIA/COPIA CLAUS/CERRAJERIA"]
ferreterias = ferreterias.drop(columns=["categoria"])
ferreterias.reset_index(drop=True, inplace=True)

electrodomesticos = actividades_economicas[actividades_economicas["categoria"] == "59###ELECTRODOMÈSTICS"]
electrodomesticos = electrodomesticos.drop(columns=["categoria"])
electrodomesticos.reset_index(drop=True, inplace=True)

muebles = actividades_economicas[actividades_economicas["categoria"] == "99###MOBLES I DECORACIÓ"]
muebles = muebles.drop(columns=["categoria"])
muebles.reset_index(drop=True, inplace=True)

tiendas_ropa = actividades_economicas[actividades_economicas["categoria"] == "121###ROBA"]
tiendas_ropa = tiendas_ropa.drop(columns=["categoria"])
tiendas_ropa.reset_index(drop=True, inplace=True)

floristerias = actividades_economicas[actividades_economicas["categoria"] == "99###MOBLES I DECORACIÓ"]
floristerias = floristerias.drop(columns=["categoria"])
floristerias.reset_index(drop=True, inplace=True)

zapaterias = actividades_economicas[actividades_economicas["categoria"] == "122###SABATERIA"]
zapaterias = zapaterias.drop(columns=["categoria"])
zapaterias.reset_index(drop=True, inplace=True)

lavanderias = actividades_economicas[actividades_economicas["categoria"] == "36###BUGADERIA"]
lavanderias = lavanderias.drop(columns=["categoria"])
lavanderias.reset_index(drop=True, inplace=True)

mecanicos = actividades_economicas[actividades_economicas["categoria"] == "133###VENDA I REPARACIÓ DE VEHÍCLES DE MOTOR"]
mecanicos = mecanicos.drop(columns=["categoria"])
mecanicos.reset_index(drop=True, inplace=True)

herboristerias = actividades_economicas[actividades_economicas["categoria"] == "51###DIETÈTICA I HERBORISTERIA"]
herboristerias = herboristerias.drop(columns=["categoria"])
herboristerias.reset_index(drop=True, inplace=True)

droguerias = actividades_economicas[actividades_economicas["categoria"] == "54###DROGUERIA - PERFUMERIA"]
droguerias = droguerias.drop(columns=["categoria"])
droguerias.reset_index(drop=True, inplace=True)

tiendas_deportivas = actividades_economicas[actividades_economicas["categoria"] == "87###JOQUINES I ARTICLES D'ESPORT"]
tiendas_deportivas = tiendas_deportivas.drop(columns=["categoria"])
tiendas_deportivas.reset_index(drop=True, inplace=True)

basureros = actividades_economicas[actividades_economicas["categoria"] == "31###BASARS"]
basureros = basureros.drop(columns=["categoria"])
basureros.reset_index(drop=True, inplace=True)

### 3.2. Ocio

#### 3.2.1 Lugares de ocio

ocio.head()

ocio["latitud"] = ocio["geometry"].y
ocio["longitud"] = ocio["geometry"].x
ocio = ocio.drop(columns=["f_id", "poi_nom", "poi_category", "poi_adreca", "poi_telefon", "poi_horari", "poi_observacions", "tipo_poi",
                          "categoria", "idioma", "atributo", "geometry"])
ocio.poi_type.value_counts()

lugares_interes = ocio[ocio["poi_type"] == "Llocs d'interès"]
lugares_interes = lugares_interes.drop(columns=["poi_type"])
lugares_interes.reset_index(drop=True, inplace=True)
lugares_interes.head()

bares = actividades_economicas[actividades_economicas["categoria"] == "30###BARS I CAFETERIES"]
bares = bares.drop(columns=["categoria"])
bares.reset_index(drop=True, inplace=True)

restaurantes = actividades_economicas[actividades_economicas["categoria"] == "120###RESTAURANTS"]
restaurantes = restaurantes.drop(columns=["categoria"])
restaurantes.reset_index(drop=True, inplace=True)

#### 3.2.2 Áreas de descanso

areas_descanso = ocio[ocio["poi_type"] == "Àrees d'esbarjo"]
areas_descanso = areas_descanso.drop(columns=["poi_type"])
areas_descanso.reset_index(drop=True, inplace=True)
areas_descanso.head()

### 3.3. Cultura

cultura.head(2)

cultura["latitud"] = cultura["geometry"].y
cultura["longitud"] = cultura["geometry"].x
cultura = cultura.drop(columns=["f_id", "poi_nom", "poi_category", "poi_adreca", "poi_telefon", "poi_web", "poi_horari", 
                                "poi_observacions", "tipo_poi", "categoria", "idioma", "atributo", "geometry"])
cultura.poi_type.value_counts()

parques = cultura[cultura["poi_type"] == "Parcs, places i jardins"]
parques = parques.drop(columns=["poi_type"])
parques.reset_index(drop=True, inplace=True)

espacios_culturales = cultura[cultura["poi_type"] == "Espais socials i culturals"]
espacios_culturales = espacios_culturales.drop(columns=["poi_type"])
espacios_culturales.reset_index(drop=True, inplace=True)

exposiciones = cultura[cultura["poi_type"] == "Exposicions i galeries d'art"]
exposiciones = exposiciones.drop(columns=["poi_type"])
exposiciones.reset_index(drop=True, inplace=True)

bibliotecas = cultura[cultura["poi_type"] == "Biblioteques i arxius"]
bibliotecas = bibliotecas.drop(columns=["poi_type"])
bibliotecas.reset_index(drop=True, inplace=True)

### 3.4. Educación

educacion.head(2)

educacion["latitud"] = educacion["geometry"].y
educacion["longitud"] = educacion["geometry"].x
educacion = educacion.drop(columns=["f_id", "poi_nom", "poi_category", "poi_adreca", "poi_telefon", "poi_web", "tipo_poi", "categoria", 
                                    "idioma", "atributo", "geometry"])
educacion.poi_type.value_counts()

colegios = educacion[educacion["poi_type"] == "Educació infantil i primària"]
colegios = colegios.drop(columns=["poi_type"])
colegios.reset_index(drop=True, inplace=True)

guarderias = educacion[educacion["poi_type"] == "Escoles bressol"]
guarderias = guarderias.drop(columns=["poi_type"])
guarderias.reset_index(drop=True, inplace=True)

institutos = educacion[educacion["poi_type"] == "Educació secundària"]
institutos = institutos.drop(columns=["poi_type"])
institutos.reset_index(drop=True, inplace=True)

otros_centros = educacion[educacion["poi_type"] == "Altres"]
otros_centros = otros_centros.drop(columns=["poi_type"])
otros_centros.reset_index(drop=True, inplace=True)

### 3.5. Sanidad

sanidad.head(2)

sanidad["latitud"] = sanidad["geometry"].y
sanidad["longitud"] = sanidad["geometry"].x
sanidad = sanidad.drop(columns=["f_id", "poi_nom", "poi_category", "poi_adreca", "poi_telefon", "poi_web", "poi_horari", 
                                "poi_observacions", "tipo_poi", "categoria", "idioma", "atributo", "geometry"])
sanidad.poi_type.value_counts()

farmacias = sanidad[sanidad["poi_type"] == "Farmàcies"]
farmacias = farmacias.drop(columns=["poi_type"])
farmacias.reset_index(drop=True, inplace=True)

residencias = sanidad[sanidad["poi_type"] == "Residències"]
residencias = residencias.drop(columns=["poi_type"])
residencias.reset_index(drop=True, inplace=True)

cap = sanidad[sanidad["poi_type"] == "CAP"]
cap = cap.drop(columns=["poi_type"])
cap.reset_index(drop=True, inplace=True)

cruz_roja = servicios[servicios["poi_nom"] == "Creu Roja"]
cruz_roja = cruz_roja.drop(columns=["poi_nom"])
cruz_roja.reset_index(drop=True, inplace=True)

### 3.6. Deporte

deporte.head(2)

deporte["latitud"] = deporte["geometry"].y
deporte["longitud"] = deporte["geometry"].x
deporte = deporte.drop(columns=["f_id", "poi_category", "poi_type", "poi_nom", "poi_adreca", "poi_telefon", "poi_web", 
                                "poi_observacions", "tipo_poi", "categoria", "idioma", "atributo", "geometry"])

### 3.7. Movilidad

movilidad.head(2)

movilidad["latitud"] = movilidad["geometry"].y
movilidad["longitud"] = movilidad["geometry"].x
movilidad = movilidad.drop(columns=["f_id", "poi_category", "poi_adreca", "poi_telefon", "poi_horari", "poi_observacions", 
                                  "tipo_poi", "categoria", "idioma", "atributo"])
movilidad.poi_type.value_counts()

autobus = movilidad[movilidad["poi_type"] == "Autobús"]
autobus = autobus.drop(columns=["poi_type"])
autobus.reset_index(drop=True, inplace=True)

taxi = movilidad[movilidad["poi_type"] == "Taxi"]
taxi = taxi.drop(columns=["poi_type"])
taxi.reset_index(drop=True, inplace=True)

transporte_demanda = movilidad[movilidad["poi_type"] == "Transport a la demanda"]
transporte_demanda = transporte_demanda.drop(columns=["poi_type"])
transporte_demanda.reset_index(drop=True, inplace=True)

### 3.8. Servicios públicos

servicios.poi_nom.value_counts()

correos = servicios[servicios["poi_nom"] == "Correus"]
correos = correos.drop(columns=["poi_nom"])
correos.reset_index(drop=True, inplace=True)

notaria = servicios[servicios["poi_nom"] == "Notaria"]
notaria = notaria.drop(columns=["poi_nom"])
notaria.reset_index(drop=True, inplace=True)

administracion.head(2)

administracion["latitud"] = administracion["geometry"].y
administracion["longitud"] = administracion["geometry"].x

administracion = administracion.drop(columns=["f_id", "poi_category", "poi_type", "poi_adreca", "poi_telefon", "poi_web", "poi_horari", 
                                              "poi_observacions", "tipo_poi", "categoria", "idioma", "atributo"])

ayuntamiento = administracion[administracion["poi_nom"] == "Ajuntament de Castellar - Palau Tolrà"]
ayuntamiento = ayuntamiento.drop(columns=["poi_nom"])
ayuntamiento.reset_index(drop=True, inplace=True)

## 4. Cálculo de distancias
En primer lugar se guardan todas las variables en un array para poder acceder a ellas más fácilmente

localizaciones = [supermercados, pescaderias, alimentacion, panaderias, carnicerias, alimentacion_no_especializada, 
                  peluquerias, mercado_municipal, mercado_ambulante, electricistas_fontaneros, ferreterias, 
                  electrodomesticos, muebles, tiendas_ropa, floristerias, lugares_interes, bares, restaurantes, 
                  areas_descanso, parques, espacios_culturales, exposiciones, bibliotecas, colegios, guarderias, institutos,
                  otros_centros, farmacias, residencias, cap, cruz_roja, deporte, autobus, taxi, transporte_demanda, 
                  correos, notaria, ayuntamiento]

a continuación se define la función que calcula la distancia mínima

# función que calcula la distancia mínima
def distancia_minima(distancia1, distancia2):
    if distancia1 < distancia2:
        distancia_min = distancia1
    else:
        distancia_min = distancia2
    return distancia_min

Se calcula la distancia euclídea utilizando la función haversine

# añadir las localizaciones al dataset de accesos, dándoles un valor inicial de 0
for i in range(0, len(localizaciones)):
    accesos[i] = 0

# cálculo de distancia euclídea
for acceso in accesos.index:
    for i in range(0, len(localizaciones)):
        # se da una distancia inicial muy alta para que esta sea superada con el primer punto
        dist_minima = 10000.0
        for localizacion in localizaciones[i].index:
            # calcula la distancia
            dist = hs.haversine((accesos.latitud[acceso], accesos.longitud[acceso]), 
                                (localizaciones[i].latitud[localizacion], localizaciones[i].longitud[localizacion])) 
            # compara la distancia y guarda la menor
            dist_minima = distancia_minima(dist_minima, dist)
            accesos[i][acceso] = dist_minima
            
accesos.head()

La distancia está en kilómetros, por lo que hay que pasarla a metros

accesos_euc = accesos.copy()

for columna in accesos.iloc[:, 2:].columns:
    accesos_euc[columna] = accesos[columna] * 1000
    
accesos_euc.to_csv('Resultados/accesos_distancia_euc.csv')

accesos_euc.head()
