==> Análisis y Cálculo de Calidad de vida/4. Cálculo de calidad de vida.ipynb <==
# Cálculo de calidad de vida
Una vez calculadas las distancias, se calculará el tiempo que se tarda en recorrerlas
## 1. Importación de librerías y establecimiento de directorio de trabajo

import os
import pandas as pd
import random
import warnings
warnings.simplefilter("ignore")

os.chdir("C:/Users/aleja/OneDrive - Universidad Politécnica de Madrid/Universidad/Asignaturas/Cuarto/Trabajo de Fin de Grado/")

## 2. Carga de datos

ubicaciones = pd.read_csv("Resultados/accesos_distancia_completo.csv", index_col="Unnamed: 0")

localizaciones = ["supermercados", "pescaderias", "alimentacion", "panaderias", "carnicerias", 
                  "alimentacion_no_especializada", "peluquerias", "mercado_municipal", "mercado_ambulante", 
                  "electricistas_fontaneros", "ferreterias", "electrodomesticos", "muebles", "tiendas_ropa", "floristerias",
                  "lugares_interes", "bares", "restaurantes", "areas_de_descanso", "parques", "espacios_culturales", 
                  "exposiciones", "bibliotecas", "colegios", "guarderias", "institutos", "otros_centros", "farmacias", 
                  "residencias", "cap", "cruz_roja", "deporte", "autobus", "taxi", "transporte_demanda", "correos", 
                  "notaria", "ayuntamiento"]

ubicaciones.head()

## 3. Cálculo de tiempo

Para calcular el tiempo que tarda cada persona en recorrer una distancia, se utilizará el estudio realizado por causadirecta para saber a qué velocidad anda cada persona según su rango de edad:

https://causadirecta.com/especial/calculo-de-velocidades/tablas/tabla-relacion-velocidad-de-peatones-caminando#:~:text=TABLA%20RELACI%C3%93N%20VELOCIDAD%20DE%20PEATONES%20CAMINANDO%20%20,%20%202%2C04%20%206%20more%20rows%20

ubicaciones_tiempo = ubicaciones.copy()

# 10 a 14
velocidad = 1.65 # m/s

for columna in ubicaciones.iloc[:, 2:].columns:
    ubicaciones_tiempo[columna] = (ubicaciones[columna] / velocidad) / 60
ubicaciones_tiempo.to_csv('Resultados/Tiempo/Accesos_Tiempo_10a14.csv')

# 15 a 19
velocidad = 1.62 # m/s

for columna in ubicaciones.iloc[:, 2:].columns:
    ubicaciones_tiempo[columna] = (ubicaciones[columna] / velocidad) / 60
ubicaciones_tiempo.to_csv('Resultados/Tiempo/Accesos_Tiempo_15a19.csv')

# 20 a 44
velocidad = 1.59 # m/s

for columna in ubicaciones.iloc[:, 2:].columns:
    ubicaciones_tiempo[columna] = (ubicaciones[columna] / velocidad) / 60
ubicaciones_tiempo.to_csv('Resultados/Tiempo/Accesos_Tiempo_20a44.csv')

# 45 a 54
velocidad = 1.5 # m/s

for columna in ubicaciones.iloc[:, 2:].columns:
    ubicaciones_tiempo[columna] = (ubicaciones[columna] / velocidad) / 60
ubicaciones_tiempo.to_csv('Resultados/Tiempo/Accesos_Tiempo_45a54.csv')

# 55 a 54
velocidad = 1.44 # m/s

for columna in ubicaciones.iloc[:, 2:].columns:
    ubicaciones_tiempo[columna] = (ubicaciones[columna] / velocidad) / 60
ubicaciones_tiempo.to_csv('Resultados/Tiempo/Accesos_Tiempo_55a64.csv')

# 65 mas
velocidad = 1.26 # m/s

for columna in ubicaciones.iloc[:, 2:].columns:
    ubicaciones_tiempo[columna] = (ubicaciones[columna] / velocidad) / 60
ubicaciones_tiempo.to_csv('Resultados/Tiempo/Accesos_Tiempo_65mas.csv')

## 4. Binarización

ubicaciones_tiempo10 = pd.read_csv("Resultados/Tiempo/Accesos_Tiempo_10a14.csv", index_col="Unnamed: 0")
ubicaciones_tiempo15 = pd.read_csv("Resultados/Tiempo/Accesos_Tiempo_15a19.csv", index_col="Unnamed: 0")
ubicaciones_tiempo20 = pd.read_csv("Resultados/Tiempo/Accesos_Tiempo_20a44.csv", index_col="Unnamed: 0")
ubicaciones_tiempo45 = pd.read_csv("Resultados/Tiempo/Accesos_Tiempo_45a54.csv", index_col="Unnamed: 0")
ubicaciones_tiempo55 = pd.read_csv("Resultados/Tiempo/Accesos_Tiempo_55a64.csv", index_col="Unnamed: 0")
ubicaciones_tiempo65 = pd.read_csv("Resultados/Tiempo/Accesos_Tiempo_65mas.csv", index_col="Unnamed: 0")

edades = [ubicaciones_tiempo10, ubicaciones_tiempo15, ubicaciones_tiempo20, ubicaciones_tiempo45, ubicaciones_tiempo55, ubicaciones_tiempo65]

for rango_edad in range(0, len(edades)):
    for localizacion in range(0, len(localizaciones)):
        for ubicacion in edades[rango_edad].index:
            if edades[rango_edad][str(localizacion)][ubicacion] < 15 and edades[rango_edad][str(localizacion)][ubicacion] != 0:
                edades[rango_edad][str(localizacion)][ubicacion] = int(1)
            else:
                edades[rango_edad][str(localizacion)][ubicacion] = int(0) 
                
edades[0].to_csv('Resultados/Binario/Accesos_Binario_10a14.csv')
edades[1].to_csv('Resultados/Binario/Accesos_Binario_15a19.csv')
edades[2].to_csv('Resultados/Binario/Accesos_Binario_20a44.csv')
edades[3].to_csv('Resultados/Binario/Accesos_Binario_45a54.csv')
edades[4].to_csv('Resultados/Binario/Accesos_Binario_55a64.csv')
edades[5].to_csv('Resultados/Binario/Accesos_Binario_65mas.csv')

## 5. Cálculo de indicadores
### 5.1. Carga de datos 
Se cargarán los datos y se añadirán los nombres de sus atributos para que puedan ser agrupados por indicadores

ubicaciones_numero10 = pd.read_csv("Resultados/Binario/Accesos_Binario_10a14.csv", index_col="Unnamed: 0")
ubicaciones_numero15 = pd.read_csv("Resultados/Binario/Accesos_Binario_15a19.csv", index_col="Unnamed: 0")
ubicaciones_numero20 = pd.read_csv("Resultados/Binario/Accesos_Binario_20a44.csv", index_col="Unnamed: 0")
ubicaciones_numero45 = pd.read_csv("Resultados/Binario/Accesos_Binario_45a54.csv", index_col="Unnamed: 0")
ubicaciones_numero55 = pd.read_csv("Resultados/Binario/Accesos_Binario_55a64.csv", index_col="Unnamed: 0")
ubicaciones_numero65 = pd.read_csv("Resultados/Binario/Accesos_Binario_65mas.csv", index_col="Unnamed: 0")

edades = [ubicaciones_numero10, ubicaciones_numero15, ubicaciones_numero20, ubicaciones_numero45, ubicaciones_numero55, ubicaciones_numero65]

for rango_edad in range(0, len(edades)):
    for localizacion in range(0, len(localizaciones)):
        edades[rango_edad] = edades[rango_edad].rename(columns={str(localizacion): str(localizaciones[localizacion])})

edades[0].head()

### 5.2. Agrupación en indicadores
Se agruparán todos los puntos en indicadores, con que un punto tenga el valor de 1, el indicador valdrá 1

indicadores = edades.copy()
for rango_edad in range(0, len(indicadores)):
    indicadores[rango_edad]["comercios_basicos"] = 0
    indicadores[rango_edad]["comercios_secundarios"] = 0
    indicadores[rango_edad]["areas_descanso"] = 0
    indicadores[rango_edad]["lugares_ocio"] = 0
    indicadores[rango_edad]["centros_educativos"] = 0
    indicadores[rango_edad]["espacios_culturales"] = 0
    indicadores[rango_edad]["farmacias"] = 0
    indicadores[rango_edad]["residencias"] = 0
    indicadores[rango_edad]["centros_sanitarios"] = 0
    indicadores[rango_edad]["centros_deportivos"] = 0
    indicadores[rango_edad]["estaciones_mobilidad"] = 0
    indicadores[rango_edad]["servicios_publicos"] = 0

    for ubicacion in edades[rango_edad].index:
        if (edades[rango_edad]["supermercados"][ubicacion] == 1 or 
            edades[rango_edad]["pescaderias"][ubicacion] == 1 or
            edades[rango_edad]["alimentacion"][ubicacion] == 1 or
            edades[rango_edad]["panaderias"][ubicacion] == 1 or
            edades[rango_edad]["carnicerias"][ubicacion] == 1 or
            edades[rango_edad]["alimentacion_no_especializada"][ubicacion] == 1 or
            edades[rango_edad]["peluquerias"][ubicacion] == 1 or
            edades[rango_edad]["mercado_municipal"][ubicacion] == 1 or
            edades[rango_edad]["mercado_ambulante"][ubicacion] == 1):
        
            indicadores[rango_edad]["comercios_basicos"][ubicacion] = 1
        
        if (edades[rango_edad]["electricistas_fontaneros"][ubicacion] == 1 or 
            edades[rango_edad]["ferreterias"][ubicacion] == 1 or 
            edades[rango_edad]["electrodomesticos"][ubicacion] == 1 or 
            edades[rango_edad]["muebles"][ubicacion] == 1 or 
            edades[rango_edad]["tiendas_ropa"][ubicacion] == 1 or 
            edades[rango_edad]["floristerias"][ubicacion] == 1):

            indicadores[rango_edad]["comercios_secundarios"][ubicacion] = 1

        if (edades[rango_edad]["areas_de_descanso"][ubicacion] == 1 or 
            edades[rango_edad]["parques"][ubicacion] == 1):

            indicadores[rango_edad]["areas_descanso"][ubicacion] = 1

        if (edades[rango_edad]["lugares_interes"][ubicacion] == 1 or 
            edades[rango_edad]["bares"][ubicacion] == 1 or 
            edades[rango_edad]["restaurantes"][ubicacion] == 1):

            indicadores[rango_edad]["lugares_ocio"][ubicacion] = 1

        if (edades[rango_edad]["colegios"][ubicacion] == 1 or 
            edades[rango_edad]["guarderias"][ubicacion] == 1 or 
            edades[rango_edad]["institutos"][ubicacion] == 1 or 
            edades[rango_edad]["otros_centros"][ubicacion] == 1):
            
            indicadores[rango_edad]["centros_educativos"][ubicacion] = 1

        if (edades[rango_edad]["espacios_culturales"][ubicacion] == 1 or 
            edades[rango_edad]["exposiciones"][ubicacion] == 1 or 
            edades[rango_edad]["bibliotecas"][ubicacion] == 1):

            indicadores[rango_edad]["espacios_culturales"][ubicacion] = 1

        if (edades[rango_edad]["farmacias"][ubicacion] == 1):

            indicadores[rango_edad]["farmacias"][ubicacion] = 1

        if (edades[rango_edad]["residencias"][ubicacion] == 1):

            indicadores[rango_edad]["residencias"][ubicacion] = 1

        if (edades[rango_edad]["cap"][ubicacion] == 1 or 
            edades[rango_edad]["cruz_roja"][ubicacion] == 1):
            
            indicadores[rango_edad]["centros_sanitarios"][ubicacion] = 1

        if (edades[rango_edad]["deporte"][ubicacion] == 1):
            
            indicadores[rango_edad]["centros_deportivos"][ubicacion] = 1

        if (edades[rango_edad]["autobus"][ubicacion] == 1 or 
            edades[rango_edad]["taxi"][ubicacion] == 1 or 
            edades[rango_edad]["transporte_demanda"][ubicacion] == 1):

            indicadores[rango_edad]["estaciones_mobilidad"][ubicacion] = 1

        if (edades[rango_edad]["correos"][ubicacion] == 1 or 
            edades[rango_edad]["notaria"][ubicacion] == 1 or 
            edades[rango_edad]["ayuntamiento"][ubicacion] == 1):

            indicadores[rango_edad]["servicios_publicos"][ubicacion] = 1

### 5.3. Limpiar y guardar nuevos datasets

for rango_edad in range(0, len(indicadores)):
    indicadores[rango_edad] = indicadores[rango_edad].drop(columns=['supermercados', 'pescaderias', 'alimentacion', 
                                                                    'panaderias', 'carnicerias', 
                                                                    'alimentacion_no_especializada', 'peluquerias', 
                                                                    'mercado_municipal', 'mercado_ambulante', 
                                                                    'electricistas_fontaneros', 'ferreterias', 
                                                                    'electrodomesticos', 'muebles', 'tiendas_ropa', 
                                                                    'floristerias', 'lugares_interes', 'bares', 
                                                                    'restaurantes', 'areas_de_descanso', 'parques', 
                                                                    'espacios_culturales', 'exposiciones', 'bibliotecas', 
                                                                    'colegios', 'guarderias', 'institutos', 'otros_centros',
                                                                    'farmacias', 'residencias', 'cap', 'cruz_roja', 
                                                                    'deporte', 'autobus','taxi', 'transporte_demanda', 
                                                                    'correos', 'notaria', 'ayuntamiento'])

indicadores[0].to_csv('Resultados/Indicadores/Accesos_Indicadores_10a14.csv')
indicadores[1].to_csv('Resultados/Indicadores/Accesos_Indicadores_15a19.csv')
indicadores[2].to_csv('Resultados/Indicadores/Accesos_Indicadores_20a44.csv')
indicadores[3].to_csv('Resultados/Indicadores/Accesos_Indicadores_45a54.csv')
indicadores[4].to_csv('Resultados/Indicadores/Accesos_Indicadores_55a64.csv')
indicadores[5].to_csv('Resultados/Indicadores/Accesos_Indicadores_65mas.csv')

## 6. Cálculo de calidad de vida
### 6.1. Carga de datos

indicadores10 = pd.read_csv("Resultados/Indicadores/Accesos_Indicadores_10a14.csv", index_col="Unnamed: 0")
indicadores15 = pd.read_csv("Resultados/Indicadores/Accesos_Indicadores_15a19.csv", index_col="Unnamed: 0")
indicadores20 = pd.read_csv("Resultados/Indicadores/Accesos_Indicadores_20a44.csv", index_col="Unnamed: 0")
indicadores45 = pd.read_csv("Resultados/Indicadores/Accesos_Indicadores_45a54.csv", index_col="Unnamed: 0")
indicadores55 = pd.read_csv("Resultados/Indicadores/Accesos_Indicadores_55a64.csv", index_col="Unnamed: 0")
indicadores65 = pd.read_csv("Resultados/Indicadores/Accesos_Indicadores_65mas.csv", index_col="Unnamed: 0")

indicadores = [indicadores10, indicadores15, indicadores20, indicadores45, indicadores55, indicadores65]

encuesta = pd.read_csv("Encuestas/Encuesta.csv")

### 6.2. Cálculo

for rango_edad in range(0, len(indicadores)):
    indicadores[rango_edad]["calidad_vida"] = 0    
    indicadores[rango_edad]["edad"] = 0      
    
    for ubicacion in indicadores[rango_edad].index:
        if rango_edad == 0:
            indicadores[rango_edad]["edad"][ubicacion] = random.randint(10, 14)
        if rango_edad == 1:
            indicadores[rango_edad]["edad"][ubicacion] = random.randint(15, 19)
        if rango_edad == 2:
            indicadores[rango_edad]["edad"][ubicacion] = random.randint(20, 44)
        if rango_edad == 3:
            indicadores[rango_edad]["edad"][ubicacion] = random.randint(45, 54)
        if rango_edad == 4:
            indicadores[rango_edad]["edad"][ubicacion] = random.randint(55, 64)
        if rango_edad == 5:
            indicadores[rango_edad]["edad"][ubicacion] = random.randint(65, 80)
            
        for indicador in indicadores[rango_edad]:
            if indicador == "latitud" or indicador == "longitud" or indicador == "calidad_vida" or indicador == "edad":
                continue
            else:
                indicadores[rango_edad]["calidad_vida"][ubicacion] = indicadores[rango_edad]["calidad_vida"][ubicacion] + (indicadores[rango_edad][indicador][ubicacion] * encuesta[indicador][0])
                
indicadores[0].to_csv('Resultados/Calidad_Vida/Accesos_Calidad_Vida_10a14.csv')
indicadores[1].to_csv('Resultados/Calidad_Vida/Accesos_Calidad_Vida_15a19.csv')
indicadores[2].to_csv('Resultados/Calidad_Vida/Accesos_Calidad_Vida_20a44.csv')
indicadores[3].to_csv('Resultados/Calidad_Vida/Accesos_Calidad_Vida_45a54.csv')
indicadores[4].to_csv('Resultados/Calidad_Vida/Accesos_Calidad_Vida_55a64.csv')
indicadores[5].to_csv('Resultados/Calidad_Vida/Accesos_Calidad_Vida_65mas.csv')          

## 7. Unificar dataset
Por último se unirán todos los datasets con calidad de vida en un único dataset

qol10 = pd.read_csv("Resultados/Calidad_Vida/Accesos_Calidad_Vida_10a14.csv", index_col="Unnamed: 0")
qol15 = pd.read_csv("Resultados/Calidad_Vida/Accesos_Calidad_Vida_15a19.csv", index_col="Unnamed: 0")
qol20 = pd.read_csv("Resultados/Calidad_Vida/Accesos_Calidad_Vida_20a44.csv", index_col="Unnamed: 0")
qol45 = pd.read_csv("Resultados/Calidad_Vida/Accesos_Calidad_Vida_45a54.csv", index_col="Unnamed: 0")
qol55 = pd.read_csv("Resultados/Calidad_Vida/Accesos_Calidad_Vida_55a64.csv", index_col="Unnamed: 0")
qol65 = pd.read_csv("Resultados/Calidad_Vida/Accesos_Calidad_Vida_65mas.csv", index_col="Unnamed: 0")

qol = pd.concat([qol10, qol15, qol20, qol45, qol55, qol65], axis=0)

qol = qol.drop(columns=["areas_descanso", "lugares_ocio", "centros_educativos", "centros_sanitarios", "centros_deportivos", 
                        "servicios_publicos", "lugares_ocio", "estaciones_mobilidad", "comercios_basicos", 
                        "comercios_secundarios"])
qol = qol.reindex(columns=['latitud','longitud','edad','calidad_vida'])
qol = qol.reset_index()
qol = qol.drop(columns=['index'])

qol.to_csv("Resultados/Accesos_Calidad_Vida.csv")

qol.head()
