==> Análisis y Cálculo de Calidad de vida/5. Preprocesamiento SHP Accesos.ipynb <==
# Paso a formato Shapefile
También se guardará el dataset en formato shp para poder visualizar el resultado en forma de mapa
## 1. Importación de librerías y establecimiento de directorio de trabajo

import os
import numpy as np
import pandas as pd
import geopandas as gpd
import warnings
warnings.simplefilter("ignore")

os.chdir("C:/Users/aleja/OneDrive - Universidad Politécnica de Madrid/Universidad/Asignaturas/Cuarto/Trabajo de Fin de Grado")

## 2. Carga de datos

accesos = pd.read_csv("Resultados/Accesos_Calidad_Vida.csv", index_col="Unnamed: 0")
accesos.head()

## 3. Guardar en formato shp
### 3.1. Todos los datos

accesos_geo = gpd.GeoDataFrame(accesos.drop(['latitud', 'longitud'], axis=1), 
                               crs="EPSG:4326", 
                               geometry=gpd.points_from_xy(accesos['longitud'], accesos['latitud']))

accesos_geo.to_file("Resultados/Accesos_Calidad_Vida.shp")

accesos_geo.head()

### 3.2. Menores de 18 años

accesos_geo_menores = accesos_geo[accesos_geo["edad"] < 18]
accesos_geo_menores = accesos_geo_menores.reset_index()
accesos_geo_menores = accesos_geo_menores.drop(columns=['index'])

accesos_menores = accesos[accesos["edad"] < 18]
accesos_menores = accesos_menores.reset_index()
accesos_menores = accesos_menores.drop(columns=['index'])

accesos_menores.to_csv("Resultados/Accesos_qol_menores.csv")
accesos_geo_menores.to_file("Resultados/Accesos_qol_menores.shp")

accesos_geo_menores.head()

### 3.3. Mayores de 18 años y menores de 65 años

accesos_geo_medianos = accesos_geo[accesos_geo["edad"] < 65]
accesos_geo_medianos = accesos_geo_medianos[accesos_geo_medianos["edad"] > 18]
accesos_geo_medianos = accesos_geo_medianos.reset_index()
accesos_geo_medianos = accesos_geo_medianos.drop(columns=['index'])

accesos_medianos = accesos[accesos["edad"] < 65]
accesos_medianos = accesos_medianos[accesos_medianos["edad"] > 18]
accesos_medianos = accesos_medianos.reset_index()
accesos_medianos = accesos_medianos.drop(columns=['index'])

accesos_medianos.to_csv("Resultados/Accesos/Accesos_qol_medianos.csv")
accesos_geo_medianos.to_file("Resultados/Accesos/Accesos_qol_medianos.shp")

accesos_geo_medianos.head()

### 3.3. Mayores de 65 años

accesos_geo_mayores = accesos_geo[accesos_geo["edad"] > 64]
accesos_geo_mayores = accesos_geo_mayores.reset_index()
accesos_geo_mayores = accesos_geo_mayores.drop(columns=['index'])

accesos_mayores = accesos[accesos["edad"] > 64]
accesos_mayores = accesos_mayores.reset_index()
accesos_mayores = accesos_mayores.drop(columns=['index'])

accesos_mayores.to_csv("Resultados/Accesos/Accesos_qol_mayores.csv")
accesos_geo_mayores.to_file("Resultados/Accesos/Accesos_qol_mayores.shp")

accesos_geo_mayores.head()
